// src/main/resources/angular/src/app/api/model/vehicleType.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: kfiertek@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { I18n } from './i18n';


export interface VehicleType { 
    vehicleTypeId: number;
    name?: string;
    i18n?: I18n;
}